Parte 2

+++ src/main/java/me/gonzager/ex/operaciones/drones/DronComercial.java
@@ -4,11 +4,11 @@ import me.gonzager.ex.operaciones.misiones.Mision;
 
 public class DronComercial extends Dron{
 
-    public DronComercial(Double autonomia, Integer procesamiento, Mision mision) {
+    public DronComercial(Integer autonomia, Integer procesamiento, Mision mision) {
         super(autonomia, procesamiento, mision);
     }
 
-    public DronComercial(Double autonomia, Integer procesamiento) {
+    public DronComercial(Integer autonomia, Integer procesamiento) {
         super(autonomia, procesamiento);
 
     }
